fix: plumber — répertoire, score fractionnaire, et les erreurs vont au log (#136)
* fix: plumber tournait dans le mauvais répertoire, et son échec cachait les autres résultats

Deux défauts signalés à l'usage sur un même scan, et le second est le plus
grave.

1. `plumber analyze` ne prend pas d'argument de chemin : il travaille sur le
   répertoire courant. En mode binaire, rien ne disait lequel, donc il analysait
   le répertoire d'où DevDesk avait été lancé et répondait « --project is
   required (could not auto-detect from git remote) » — un échec à propos d'un
   dépôt que personne ne lui avait demandé de regarder. Le conteneur avait déjà
   son -w ; le binaire n'avait rien.

   toolCmd porte donc un Dir, et String() le rend (« cd <dir> && … ») : une
   invocation journalisée qu'on ne peut pas rejouer dit moins qu'elle n'en a
   l'air. Trivy et Gitleaks prennent leur cible dans argv et ne posent rien.

   Mesuré sur le dépôt signalé : l'erreur passe de « --project is required » à
   « GITLAB_TOKEN is required », c'est-à-dire plumber ayant trouvé le projet et
   réclamant la suite.

2. La vue des résultats remplaçait toute la table par le panneau d'avertissements
   dès que Result.Errors n'était pas vide, au motif écrit qu'un scan échoué n'a
   pas de résultats partiels à parcourir. C'était vrai tant qu'une erreur voulait
   dire que rien n'avait tourné ; ça a cessé de l'être le jour où une étape peut
   échouer à côté de trois qui réussissent. L'échec de plumber emportait donc
   tous les CVE et tous les secrets du même scan — des findings qui existent et
   qu'on ne voit pas, la pire façon pour un résultat d'être faux.

   Le panneau ne remplace la table que si TotalFindings() vaut zéro ; sinon il
   se place au-dessus, sur tous les onglets, parce que l'échec porte sur le scan
   et non sur l'onglet qu'on lit. showsResultTabs suit la même règle, sans quoi
   la barre d'onglets disparaîtrait d'un écran qui en a besoin.

   La hauteur du bandeau est mesurée sur le texte rendu et non comptée à la
   main : le panneau se replie, donc sa hauteur dépend de la largeur du terminal
   et de ce que l'outil avait à dire.

Entire-Checkpoint: 94e66e96541f

* fix: un score fractionnaire, et les erreurs vont au log plutôt qu'à l'écran

Deux choses, signalées à l'usage sur le même scan.

1. finalPoints est un **flottant**. Les deux fixtures notaient un nombre entier,
   donc un int les décodait, et le premier dépôt à obtenir un score fractionnaire
   faisait échouer toute l'étape :

     cannot unmarshal number 25.698320532936123 into Go struct field
     plumberScore.finalPoints of type int

   La bannière de plumber affiche une décimale, ce qui le disait déjà. Le score
   est un flottant de bout en bout et la ligne de tête en affiche une décimale
   elle aussi — arrondir à l'entier effacerait la différence entre deux dépôts.

2. Une erreur ne s'affiche plus dans le TUI. Elle va au log, et le footer dit
   d'aller y voir (Rule 128).

   Le panneau d'avertissements remplaçait toute la table dès que Result.Errors
   n'était pas vide ; le replier en bandeau au-dessus ne faisait que déplacer le
   problème sur les cinq onglets à la fois, en permanence, pour un message lu
   une fois. Et le stderr d'un outil remis en forme pour un viewport est
   exactement ce à quoi sert un log.

   Il ne reste plus rien qui remplace la table : elle est ce que l'état
   « résultats » affiche, toujours, et le footer porte un **status** — pas un
   message — qui nomme les étapes en échec et renvoie au log. Un status parce
   que c'est un état du résultat : un message expire au bout de trois secondes,
   et l'utilisateur resterait devant un résultat qui a l'air complet.

   scan.recordStageError est ce qui rend ça honnête. Chaque étape appendait à
   Result.Errors et ne journalisait rien, donc la seule copie du motif était à
   l'écran, dans le panneau qui cachait les findings. Elle journalise maintenant
   au moment où elle enregistre, aux six endroits.

internal/ui/security/warnings.go disparaît avec ses trois fonctions de mise en
forme du stderr.

Entire-Checkpoint: ced3ca3ba1d7

---------

+++ internal/scan/plumber_parse.go
@@ -27,8 +27,14 @@ type plumberDoc struct {
 }
 
 type plumberScore struct {
-	Score       string            `json:"score"`
-	FinalPoints int               `json:"finalPoints"`
+	Score string `json:"score"`
+	// FinalPoints is a **float**, and the fixtures hid it: both happened to
+	// score a whole number, so an int decoded them and the first repository
+	// with a fractional score failed the whole stage with
+	// `cannot unmarshal number 25.698320532936123 into Go struct field
+	// plumberScore.finalPoints of type int`. plumber's own banner prints one
+	// decimal, which is what gave it away in hindsight.
+	FinalPoints float64           `json:"finalPoints"`
 	CodeLosses  []plumberCodeLoss `json:"codeLosses"`
 	Counts      map[string]int    `json:"counts"`
 	Losses      []json.RawMessage `json:"losses"`
